0.2б Создано динамическое изменение процента, при регистрации создается таблица пользователя для логирования изменений баллов, настроено добавление баллов с логированием в основную и личную таблицу.

+++ Math_procent.py
@@ -2,4 +2,4 @@
 
 
 def points_value(summ, procent):
-    return summ * procent / 100
+    return int(summ * procent / 100)
